Conclusion of this class with your test

+++ model/DAO/UserDAO.php
@@ -2,14 +2,19 @@
 
 namespace model\dao;
 
-require_once '../connection/Connection.php';
+require_once  '../../vendor/autoload.php';
+
+use model\connection\Connection;
+use model\interfaces\DAOInterface;
+use model\pojo\Pojo;
+use model\pojo\User;
 
 /**
  * 
  * @author vitor.brangioni
  *
  */
-class UserDAO
+class UserDAO implements DAOInterface
 {
 	private static $instance;
 	private static $conn;
@@ -17,6 +22,7 @@ class UserDAO
 	private function __construct() {
 	}
 	
+	// @DONE
 	public static function getInstance() : UserDAO
 	{
 		if (!isset(self::$instance) && !isset(self::$conn)) {
@@ -26,6 +32,7 @@ class UserDAO
 		return self::$instance;
 	}
 	
+	// @DONE
 	public function listAll()
 	{
 		try {
@@ -33,19 +40,78 @@ class UserDAO
 			$stmt = self::$conn->prepare($sql);
 			$stmt->execute();
 			
-			return $stmt->fetchAll(PDO::FETCH_ASSOC);
+			return $stmt->fetchAll(\PDO::FETCH_ASSOC);
 			
-		} catch (PDOStatement $e) {
+		} catch (\PDOStatement $e) {
 			echo $e->errorCode();	
 		}
 	}
 	
-	public function insert(User $user)
+	// @DONE
+	public function findById($id)
 	{
 		try {
-			$sql = "INSERT INTO user (username, password) VALUES (:username, :password);";
+			$sql = "SELECT * FROM user WHERE id = :id";
+			$stmt = self::$conn->prepare($sql);
+			$stmt->bindValue(":id", $id);
+			$stmt->execute();
 			
-			self::$conn->prepare($sql);
+			return $stmt->fetch(\PDO::FETCH_ASSOC);
+			
+		} catch (\PDOStatement $e) {
+			echo $e->errorCode();
+		}
+	}
+	
+	//@DONE
+	/**
+	 *
+	 * @param Enum $fiel
+	 * @param unknown $value
+	 * @return unknown
+	 */
+	public function findGeneric($fiel, $value)
+	{
+		try {
+			$sql = "SELECT * FROM user WHERE ".$fiel." = :value";
+			$stmt = self::$conn->prepare($sql);
+			$stmt->bindValue(":value", $value);
+			$stmt->execute();
+			
+			return $stmt->fetch(\PDO::FETCH_ASSOC);
+			
+		} catch (\PDOStatement $e) {
+			echo $e->errorCode();
+		}
+	}
+	
+	// @TODO
+	public function findByObject(User $user)
+	{
+		try {
+			$sql = "SELECT * FROM doctor WHERE id = :id";
+			$stmt = self::$conn->prepare($sql);
+			$stmt->bindValue(":id", $user->getId());
+			$stmt->execute();
+			
+			return $stmt->fetch(\PDO::FETCH_ASSOC);
+			
+		} catch (\PDOStatement $e) {
+			echo $e->errorCode();
+		}
+	}
+	
+	// @DONE
+	public function insert(Pojo $user)
+	{
+		if (!$user instanceof User) {
+			throw new \InvalidArgumentException();
+		}
+		
+		try {
+			$sql = "INSERT INTO user (id, username, password) VALUES (null, :username, :password)";
+			
+			$stmt = self::$conn->prepare($sql);
 			$stmt->bindValue(":username", $user->getUsername());
 			$stmt->bindValue(":password", $user->getPassword());
 			$stmt->execute();
@@ -55,8 +121,13 @@ class UserDAO
 		}
 	}
 	
-	public function edit(User $doctor)
+	// @TODO
+	public function edit(Pojo $user)
 	{
+		if (!$user instanceof User) {
+			throw new \InvalidArgumentException();
+		}
+		
 		try {
 			$sql = "UPDATE user SET username = ':username', password = ':password' WHERE id = :id";
 			
@@ -66,65 +137,18 @@ class UserDAO
 			$stmt->bindValue(":id", $user->getId());
 			$stmt->execute();
 			
-		} catch (PDOStatement $e) {
+		} catch (\PDOStatement $e) {
 			echo $e->errorCode();
 		}
 	}
 	
-	public function findById($id)
+	// @DONE
+	public function delete(Pojo $user)
 	{
-		try {
-			$sql = "SELECT * FROM user WHERE id = :id";
-			$stmt = self::$conn->prepare($sql);
-			$stmt->bindValue(":id", $id);
-			$stmt->execute();
-			
-			return $stmt->fetch(PDO::FETCH_ASSOC);
-			
-		} catch (PDOStatement $e) {
-			echo $e->errorCode();
+		if (!$user instanceof User) {
+			throw new \InvalidArgumentException();
 		}
-	}
-	
-	/**
-	 * 
-	 * @param Enum $fiel
-	 * @param unknown $value
-	 * @return unknown
-	 */
-	public function findGeneric(Enum $fiel, $value)
-	{
-		try {
-			$sql = "SELECT * FROM user WHERE :field = :value";
-			$stmt = self::$conn->prepare($sql);
-			$stmt->bindValue(":field", $fiel);
-			$stmt->bindValue(":value", $value);
-			$stmt->execute();
-			
-			return $stmt->fetch(PDO::FETCH_ASSOC);
-			
-		} catch (PDOStatement $e) {
-			echo $e->errorCode();
-		}
-	}
-	
-	public function findByObject(User $user)
-	{
-		try {
-			$sql = "SELECT * FROM doctor WHERE id = :id";
-			$stmt = self::$conn->prepare($sql);
-			$stmt->bindValue(":id", $user->getId());
-			$stmt->execute();
-			
-			return $stmt->fetch(PDO::FETCH_ASSOC);
-			
-		} catch (PDOStatement $e) {
-			echo $e->errorCode();
-		}
-	}
-	
-	public function delete(User $user)
-	{
+		
 		try {
 			$sql = "DELETE FROM user WHERE id = :id";
 			
@@ -132,26 +156,13 @@ class UserDAO
 			$stmt->bindValue(":id", $user->getId());
 			$stmt->execute();
 			
-		} catch (PDOStatement $e) {
+		} catch (\PDOStatement $e) {
 			echo $e->errorCode();
 		}
 	}
 	
+	// @TODO ?
 	private function populate($row)
 	{
-		
 	}
 }
-
-$dao = UserDAO::getInstance();
-$result = $dao->listAll();
-
-if ($result == null){
-	echo 'fail';
-}
-
- foreach ($result as $data) {
- echo $data['username'];
- }
- 
-//echo $dao->findById(1)['nome'];
